+++ idempotence4j-benchmarks/src/main/java/com/transferwise/idempotence4j/benchmarks/domain/model/OperationRepository.java
@@ -1,0 +1,5 @@
+package com.transferwise.idempotence4j.benchmarks.domain.model;
+
+public interface OperationRepository {
+    void save(Operation operation);
+}