+++ idempotence4j-benchmarks/src/main/resources/db/migration/V20201001120500__add_operation.sql
@@ -1,0 +1,6 @@
+CREATE TABLE operation (
+    id              TEXT        NOT NULL,
+    created_at      TIMESTAMP   NOT NULL,
+
+    PRIMARY KEY(id)
+);