+++ idempotence4j-postgres/src/test-integration/groovy/com/transferwise/idempotence4j/postgres/IntegrationTest.groovy
@@ -1,6 +1,7 @@
 package com.transferwise.idempotence4j.postgres
 
 import com.transferwise.idempotence4j.utils.DbUtils
+import com.transferwise.idempotence4j.utils.PropertiesLoader
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
 import groovy.sql.Sql